requirements.txt
click==8.0.3
cycler==0.11.0
dataclasses==0.8
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.4.60
Pillow==8.4.0
pkg_resources==0.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
six==1.16.0
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
